backend/requirements.txt
aniso8601==10.0.0
apispec==6.8.1
apispec-webframeworks==1.2.0
attrs==25.1.0
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.0
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
swagger-ui-py==23.9.23
tornado==6.4.2
typing_extensions==4.12.2
Werkzeug==3.1.3
